+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import Game from './Game'
+import Game from './components/Game'
 
 
 ReactDOM.render(<Game/>, document.getElementById("root"))
